test(storage): isolate idb fault controls
Extract the durable worker core and place IndexedDB fault injection only in a Vitest-resolved test entry, with a production artifact symbol audit.
