+++ gossip-sdk/package.json
@@ -30,6 +30,7 @@
     "prebuild": "rimraf dist",
     "build": "tsc && npm run copy:wasm",
     "copy:wasm": "mkdir -p dist/assets/generated/wasm dist/assets/generated/wasm-secureStorage && cp -R src/assets/generated/wasm/. dist/assets/generated/wasm/ && cp -R src/assets/generated/wasm-secureStorage/. dist/assets/generated/wasm-secureStorage/",
+    "audit:production-worker": "node scripts/audit-production-worker.mjs",
     "test": "vitest",
     "test:run": "vitest run",
     "test:coverage": "vitest run --coverage",
